Docs: Add VSCode editing guide, chore: Update dependencies
Adds instructions for editing files directly in VSCode to the README.
This includes a check for proper GitHub repository connection.

Updates various project dependencies to their latest versions,
as reflected in `package-lock.json`. Notable changes include
updates to Babel packages and the inclusion of `@esbuild/linux-x64`
as an optional dependency for Vite.